Move StaticFieldRewriter to shared file
